Map a scaling in the encoder positions to a change in Z. Enables using zPIE with positioning correction.
Note: At the moment, in this naive implementation, we require a library based on JAX. This can be changed by implementing adam, it's not hard but as it's an edge case at the moment it's in there for now. The import is only called when you're running the update to z, so normal operation is possible without installing jax.